Python GPU online reconstruction project: Rolled back to commit 074daae:
This concerns mostly function make_hitpairs_in_station;
in 074daae and cebc02e, this function does essentially the same thing;
the code in cebc02e returns an array of the right size (but actually not the right "shape") and takes 13s to process all the events;
the code in 074daae returns an oversize array (of the right "shape" i.e. array of size "number_of_pairs:2") and only takes about 2 seconds to process all the events;

I have improved the function in 074daae so that it returns an array of the right size:
After filling the hitpairs array, I just declare another array of the right size and fill it with the relevent info from the oversized array; this does not increase the processing time noticeably;
the program, as is, still takes about 2 seconds to process all events.

Next step is to have the hit pair arrays to be transmitted to function reco_tracklet_in_stations without overheads.